Refactor code to handle user context and clean up JSX layout
Refactored how we handle the user context in the app. Primarily in the
LogIn, HomePage, and Register components. This is designed to make it
easier to manage user state throughout the app. In addition, there was
unnecessary complexity in the JSX that's been cleaned up for readability
and simplicity. Reduction of some divs and improving the overall layout of the components.
The changes should make these components cleaner and easier to work with in future.