Update LikeButton functionality
Added functionality to LikeButton component that sends a post request to the server when the button is clicked. This allows connection between user's wishlist and the database, specifically triggering an update on the 'wishlist' array in the user database when the 'like' button is clicked. This is necessary so the application knows and can display what items a user has 'liked'. A context hook is used to provide user data throughout the application. The affected files are LikeButton.jsx and item-details.jsx.