Fix typo in import statement for trashcan image
A typo in the import statement for the trashcan image in the Wishlist page was corrected. The file name was spelled as 'trashCan.svg' but it has been corrected to 'trashcan.svg'. This typo prevented the correct rendering of the trashcan image, breaking the Wishlist page's functionality to delete items from the list.