Update product filtering by category ID
Category filtering in product-handler.js and item-list.jsx have been updated to use category ID instead of name. This change is necessary for the following reasons: In the product-handler.js file, Category.findOne now uses an _id field to get the category details. This is essential because categories are identified by unique ID, reducing the possibility of error in fetching the right element. In item-list.jsx, category selection now uses the category ID fetched from the URL parameters. This allows for a more accurate filtering of products according to selected categories in the front end.