Update cart functionality to support quantity and store state
This commit modifies the cart functionality to now take into account product quantity, and it also updates how the cart state is managed within the user state. The API route for adding a product to the cart now accepts a 'quantity' parameter.

To reflect these backend changes, the cart related components were also updated. This includes modifications to update the user state's cart using the server response and visually display the item count on the cart icon. Finally, populating product information when retrieving the user data was added to ensure all needed product data is available for the frontend cart display.