Update LogoGreen .svg file in the pics folder
SVG file for the logo used throughout the system has been updated in this commit. The new version of the SVG file includes a revamp of the overall logo design including shape, size and color. The changes were made due to the re-branding decision made by the marketing team for a more visually appealing look.