Update key value in cart item mapping
Modified the key value in the map function for cartItems present in cart-Page.jsx. The previous key prop was causing a warning in the console when two identical items were added to the cart, due to duplicate keys. To prevent this, the key of each cartItem was adjusted to contain a string concatenation of the product id and its index, ensuring unique keys for every render.