Update CSS for layout enhancements
Added padding to the bottom of global.scss to improve clarity and spacing for readability. The NavBar now has a fixed position at the bottom of the visible window for better mobile user interface, ensuring consistent accessibility of the navigation regardless of scroll position. The margins were adjusted to center the Navbar and accommodate for its new position.